Fix the remaining confirmed bugs
**Unbounded sizes (2).** `--stars 500000000` asked for 20 GB and
`--size 60000x60000` for 86 GB. A failed allocation aborts the process
outright — no unwind, no Drop, no panic hook — and the RawGuard was
installed before Flight::new ran, so interactively it left the user in raw
mode on the alternate screen with an invisible cursor and no prompt.

--stars now caps at 1,000,000: fifty times the automatic maximum, 40 MB of
pool. --size caps at 10,000 per axis and 2,000,000 cells, checked as a
product because each axis can be reasonable while the area is not; at about
54 bytes per cell across the four buffers a frame needs, that is 110 MB.
Both now fail as clap errors naming the limit. A terminal that reports
something preposterous is clamped rather than believed, since --size is
vetted at parse time but an ioctl answer is not. And the Flight is built
before the terminal is taken over, so bulk allocation can no longer fail
with the terminal already reconfigured.

**f32 clock (3).** Flight::time accumulated dt in f32 and stopped advancing
altogether at t = 524288 s — six days — freezing the twinkle and the shake,
which is exactly the state a tmux screensaver is left in. The phase went
coarse long before that: (time * 31.0).sin() at one hour evaluates sine at
111600, where the f32 ulp is 0.008.

The clock and the autopilot's elapsed are f64 now. The four shake terms are
evaluated in f64, where libm's argument reduction stays exact indefinitely;
the per-star twinkle phase is folded into a single turn once per frame so
the hot path keeps its f32 sin. Wrapping the clock itself was the other
option and was rejected: the shake frequencies are deliberately
incommensurate, so no period exists that would not put a visible
discontinuity in the field.

**Wrong doc comment (5).** DepthRule::Anywhere claimed to be used when
refilling the sides mid-turn; update always respawns with FarPlane, and
Anywhere is reached only from new and resize_pool.

Frame cost is unchanged at 4.4 ms, a seeded headless flight is still
byte-reproducible, and the rendered image is identical.

+++ src/render.rs
@@ -61,22 +61,24 @@ impl Renderer {
     }
 
     /// Build the camera for this instant, including shake and bank.
-    pub fn camera(&self, ship: &Ship, time: f32) -> Camera {
+    pub fn camera(&self, ship: &Ship, time: f64) -> Camera {
         let (w, h) = self.canvas.dims();
         let mut cam = Camera::new(w, h);
         if ship.shake > 0.0 {
             // Two incommensurate frequencies so the wobble never looks like a
-            // clean sine, without needing a noise table.
+            // clean sine, without needing a noise table. Evaluated in `f64`:
+            // there are only four of them per frame, and it means the argument
+            // reduction stays exact however long the process has been up.
             let amp = h as f32 * SHAKE_AMPLITUDE * ship.shake;
-            cam.cx += amp * ((time * 31.0).sin() + (time * 17.3).sin() * 0.6);
-            cam.cy += amp * ((time * 27.7).cos() + (time * 13.1).sin() * 0.6);
+            cam.cx += amp * ((time * 31.0).sin() + (time * 17.3).sin() * 0.6) as f32;
+            cam.cy += amp * ((time * 27.7).cos() + (time * 13.1).sin() * 0.6) as f32;
         }
         cam.bank = ship.bank;
         cam
     }
 
     /// Draw one frame into the cell grid. Nothing reaches the terminal yet.
-    pub fn render(&mut self, field: &StarField, ship: &Ship, cam: &Camera, time: f32, hud: &Readout) {
+    pub fn render(&mut self, field: &StarField, ship: &Ship, cam: &Camera, time: f64, hud: &Readout) {
         let warp = ship.warp_intensity();
 
         self.canvas.clear();
@@ -257,12 +259,12 @@ mod tests {
             renderer.resize(cols, rows);
             let (w, h) = renderer.canvas_dims();
             assert_eq!((w, h), (cols.max(1), (rows * 2).max(1)));
-            let cam = renderer.camera(&ship, i as f32);
+            let cam = renderer.camera(&ship, i as f64);
             field.retarget(&cam);
             for _ in 0..30 {
                 ship.update(1.0 / 60.0);
                 field.update(1.0 / 60.0, ship.speed, 0.2, 0.0, &cam);
-                renderer.render(&field, &ship, &cam, i as f32, &readout(&ship));
+                renderer.render(&field, &ship, &cam, i as f64, &readout(&ship));
             }
             renderer.present(&mut Vec::new()).unwrap();
         }
